Fix: recovery code undefined

+++ functions/src/index.ts
@@ -869,10 +869,15 @@ const IsCorrectVaultPin = async (pin : string, userId : string) : Promise<boolea
 
     if (!correct)
     {
-        correct = await bcrypt.compare(pin, (<FirebaseFirestore.DocumentData>vaultConfig.data()).recoveryCode);
+        const recoveryCode = (<FirebaseFirestore.DocumentData>vaultConfig.data()).recoveryCode;
 
-        if (correct)
-            await vaultConfig.ref.update("recoveryCode", admin.firestore.FieldValue.delete()); // Delete recovery code if the user entered it (single use code)
+        if (recoveryCode)
+        {
+            correct = await bcrypt.compare(pin, recoveryCode);
+
+            if (correct)
+                await vaultConfig.ref.update("recoveryCode", admin.firestore.FieldValue.delete()); // Delete recovery code if the user entered it (single use code)
+        }
     }
 
     return correct;
